+++ attendance-api/src/routes/attendance/beneficiaryList/removeById.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from "express";
 import { Beneficiary } from "../../../database/beneficiary";
 import { logger } from "../../../logger";
-import { RemoveBeneficiaryResponse } from "../../../types/beneficiary";
+import { BeneficiaryModel, RemoveBeneficiaryResponse } from "../../../types/beneficiary";
 import commonResponseError from "../../../util/commonResponseError";
 
 export async function removeById(req: Request, res: Response) {
@@ -9,9 +9,16 @@ export async function removeById(req: Request, res: Response) {
 	logger.info(`deleting beneficiary ${id}...`)
 
 	try {
-		const result = await Beneficiary.removeById(parseInt(id))
+		const temp: any = (await Beneficiary.fetchById(parseInt(id)))[0]
+		if (!temp) throw new Error('no beneficiary found')
+		const beneficiary: BeneficiaryModel = temp.dataValues
+
+		beneficiary.deleted = true
+
+		await Beneficiary.updateById(parseInt(id), beneficiary)
+
 		const response: RemoveBeneficiaryResponse = {
-			deletedCount: result
+			deletedCount: 1
 		}
 		logger.info('delete complete')
 		res.status(202).json(response)